Audit collection docs snippets
Complete the doc collection-example audit by normalizing snippet

surface forms and adding doc-linked compile-run coverage.

Include minimal compile-blocker fixes discovered during the

baseline run so the tree is consistent for subsequent TODO work.

+++ src/ir_lowerer/IrLowererFlowHelpers.cpp
@@ -897,7 +897,7 @@ VectorStatementHelperEmitResult tryEmitVectorStatementHelper(
 
     emitReallocateVectorData(desiredLocal, emitVectorCapacityExceeded);
 
-    instructions.push_back({IrOpcode::Jump, static_cast<int32_t>(pushBodyIndex)});
+    instructions.push_back({IrOpcode::Jump, static_cast<uint64_t>(pushBodyIndex)});
 
     const size_t errorIndex = instructions.size();
     instructions[jumpOom].imm = static_cast<int32_t>(errorIndex);
